code refactorise
