+++ src/views/customer/DataTable.tsx
@@ -1,0 +1,192 @@
+import {
+    DataGrid,
+    GridRowSelectionModel,
+    GridColDef,
+    GridActionsCellItem,
+    
+  } from "@mui/x-data-grid";
+import AddIcon from '@mui/icons-material/Add';
+import React from 'react';
+import {useEffect, useState} from "react";
+import Button from '@mui/material/Button';
+import EditIcon from "@mui/icons-material/Edit";
+import DeleteIcon from "@mui/icons-material/DeleteOutlined";
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+import CustomerForm from '../../components/customer/CustomerForm.tsx';
+import AlertDeleteCustomer from '../../components/customer/AlertDeleteCustomer.tsx';
+import { findAllCustomer } from "../../services/customer/Api.ts";
+import CustomerModel from "../../models/customer/CustomerModel.ts";
+
+
+const DataTable = () => {
+
+    const columns: GridColDef[] = [
+        { field: "customerName", 
+          headerName: "Customer Name", 
+          width: 200,
+          align: 'left',
+          headerAlign: 'left', 
+        },
+        { field: "accountNumber", 
+          headerName: "Account Number", 
+          width: 150,
+          align: 'left',
+          headerAlign: 'left', 
+        },
+        {
+          field: "state",
+          headerName: "State",
+          width: 170,
+          align: 'left',
+          headerAlign: 'left',
+        },
+        {
+          field: "city",
+          headerName: "City",
+          width: 170,
+          align: 'left',
+          headerAlign: 'left',
+        },
+        {
+          field: "zipCode",
+          headerName: "Zip Code",
+          width: 170,
+          align: 'left',
+          headerAlign: 'left',
+        },
+        {
+          field: "agency",
+          headerName: "Agency",
+          width: 170,
+          align: 'left',
+          headerAlign: 'left',
+        },
+        {
+          field: "tmcClientNumber",
+          headerName: "TMC Client Number",
+          width: 200,
+          align: 'left',
+          headerAlign: 'left',
+        },
+        {
+          field: "actions",
+          type: "actions",
+          headerName: "Actions",
+          width: 100,
+          cellClassName: "actions",
+          getActions: ({id}) => {
+            return [
+              <GridActionsCellItem
+                icon={<EditIcon />}
+                label="Edit"
+                className="textPrimary"
+                onClick={displayModalUpdate(id)}
+                color="inherit"
+              />,
+              <GridActionsCellItem
+                icon={<DeleteIcon />}
+                label="Delete"
+                onClick={deleteCustomer(id)}
+                color="inherit"
+              />
+            ];
+          }
+        }
+      ];
+    
+    
+      const [rows, setRows] = useState<CustomerModel[]>([])
+    
+      const [rowSelectionModel, setRowSelectionModel] = useState<
+        GridRowSelectionModel
+      >([]);
+      const [relaodData, setRelaodData] = useState(true)
+      const [openModal, setOpenModal] = useState(false);
+      const [openModalDelete, setOpenModalDelete] = useState(false);
+      // const [displayDeleteAll, setDisplayDeleteAll] = useState(false);
+      const [idCustomerToDel, setIdCustomerToDel] = useState();
+      const [customer, setCustomer] = useState<any>(null);
+    
+      const displayModal = () => {
+    
+          setOpenModal(true);
+        
+      }
+    
+      const displayModalUpdate = (id: any) => () => {
+        
+        const cust = rows.filter((row) => row.id === parseInt(id))[0]
+        setCustomer(cust)
+        setOpenModal(true);
+        
+      }
+    
+      const deleteCustomer = (id?: any) => () => {
+    
+        setIdCustomerToDel(id)
+        setOpenModalDelete(true)
+      };
+    
+      const onNotifmodal = (msg : boolean)=>{
+        setOpenModal(msg)
+        setOpenModalDelete(msg)
+        setCustomer(null)
+      }
+    
+      const msgSuccess = (msg: string) => {
+        toast.success(msg,
+                        {position: toast.POSITION.TOP_CENTER})
+        setRelaodData(!relaodData)
+      }
+    
+      useEffect(() => {
+        //console.log('findAllCustomer() => ', findAllCustomer().then(data => setRows(data)).catch(err => console.log('err in fecthing data', err)));
+        findAllCustomer().then(data => setRows(data))
+        
+      }, [relaodData]);
+    
+    
+      return (
+        <div className="p-4">
+            <ToastContainer />
+                  {openModal && (
+                <div>
+                  <CustomerForm onNotifmodal={onNotifmodal} customer={customer} msgSuccess={msgSuccess} rows={rows}/>
+                </div>
+              )}
+    
+                {openModalDelete && (
+                  <div>
+                    <AlertDeleteCustomer onNotifmodal={onNotifmodal} idCustomer = {idCustomerToDel} msgSuccess={msgSuccess}/>
+                  </div>
+                )}
+    
+              <h3 className="flex justify-center text-2x font-bold mb-10 mt-20">CUSTOMERS LIST</h3>
+              <Button color="primary" variant= "outlined" startIcon={<AddIcon />} onClick={displayModal}>
+                Add Customer
+              </Button> <span className='mr-3'></span>
+              <div className="mt-2 " style={{ height: 370, width: '100%' }}>
+                <DataGrid
+                  rows={rows}
+                  columns={columns}
+                  editMode="row"
+                  checkboxSelection
+                  onRowSelectionModelChange={(newRowSelectionModel) => {
+                    setRowSelectionModel(newRowSelectionModel);
+                  }}
+                  rowSelectionModel={rowSelectionModel}
+                  initialState={{
+                    pagination: {
+                      paginationModel: {  page: 0, pageSize: 5 }
+                    }
+                  }}
+                  pageSizeOptions={[5, 10, 20]}
+                />
+              </div>     
+        </div>
+      );
+}
+
+
+export default DataTable;
